feat(interaction): First step toward editable Activity via table.
