+++ web/quickplan/src/components/NodeTools/NodeTools.tsx
@@ -1,62 +1,32 @@
 import React, {FC, useEffect, useState} from 'react'
-import {Button, Divider, Drawer, Space, Table, Typography} from "antd";
+import {Button, Divider, Drawer, Space} from "antd";
 import {useStore} from "../../store/store";
 import {CpmNodeType} from "../ReactFlow/CpmTaskNode";
-import {LeftOutlined, PlusOutlined, RightOutlined} from "@ant-design/icons";
+import {LeftOutlined, RightOutlined} from "@ant-design/icons";
 import {ActivityAddButton} from "./ActivityAddButton";
 import {Edge} from "react-flow-renderer";
+import {NodesTable} from "./NodesTable";
 
-export const NodesTable: FC = () => {
-    const {nodes} = useStore()
+export const ActionsBar: FC = () => {
+    const {activeChartId, activeEdgeId, activeNodeId, edges, nodes} = useStore()
+    const [edge, edgeSet] = useState<null | Edge>(null)
+    const [node, nodeSet] = useState<null | CpmNodeType>(null)
 
-    const columns = [
-        {
-            title: 'Id',
-            dataIndex: 'id',
-            key: 'id',
-        },
-        {
-            title: 'Title',
-            dataIndex: ['data', 'label'],
-            key: 'title',
-        },
-        {
-            title: 'Duration',
-            dataIndex: ['data', 'cpm', 'duration'],
-            key: 'cpm.duration',
-        },
-        {
-            title: 'Duration Low',
-            dataIndex: ['data', 'cpm', 'duration'],
-            key: 'cpm.durationLow',
-        },
-        {
-            title: 'Duration Likely',
-            dataIndex: ['data', 'cpm', 'duration'],
-            key: 'cpm.durationLikely',
-        },
-        {
-            title: 'Duration High',
-            dataIndex: ['data', 'cpm', 'duration'],
-            key: 'cpm.durationHigh',
-        },
-        {
-            title: 'Earliest Start',
-            dataIndex: ['data', 'cpm', 'earlyStart'],
-            key: 'cpm.earliestStart',
-            sorter: (a: CpmNodeType, b: CpmNodeType) => a.data.cpm.earlyStart - b.data.cpm.earlyStart
-        },
-    ]
+    useEffect(() => {
+        const activeE = edges.filter((e) => e.id === activeEdgeId)
+        if (activeE.length === 0) {
+            edgeSet(null)
+        }
+        edgeSet(activeE[0])
+    }, [activeEdgeId, edges])
 
-    return <Table dataSource={nodes} columns={columns} size="small"/>
-}
-
-type ActionsBarProps = {
-    node: null | CpmNodeType
-    edge: null | Edge
-}
-export const ActionsBar: FC<ActionsBarProps> = ({node, edge}) => {
-    const {activeChartId} = useStore()
+    useEffect(() => {
+        const activeN = nodes.filter((n) => n.id === activeNodeId)
+        if (activeN.length === 0) {
+            nodeSet(null)
+        }
+        nodeSet(activeN[0])
+    }, [activeNodeId, nodes])
 
     if (!activeChartId) {
         return <></>
@@ -83,25 +53,7 @@ export const ActionsBar: FC<ActionsBarProps> = ({node, edge}) => {
 }
 
 export const NodeTools: FC = () => {
-    const {nodeToolsVisible, activeEdgeId, activeNodeId, edges, nodes} = useStore()
-    const [edge, edgeSet] = useState<null | Edge>(null)
-    const [node, nodeSet] = useState<null | CpmNodeType>(null)
-
-    useEffect(() => {
-        const activeE = edges.filter((e) => e.id === activeEdgeId)
-        if (activeE.length === 0) {
-            edgeSet(null)
-        }
-        edgeSet(activeE[0])
-    }, [activeEdgeId, edges])
-
-    useEffect(() => {
-        const activeN = nodes.filter((n) => n.id === activeNodeId)
-        if (activeN.length === 0) {
-            nodeSet(null)
-        }
-        nodeSet(activeN[0])
-    }, [activeNodeId, nodes])
+    const {nodeToolsVisible} = useStore()
 
     return (
         <Drawer
@@ -112,7 +64,7 @@ export const NodeTools: FC = () => {
             placement="bottom"
             visible={nodeToolsVisible}
         >
-            <ActionsBar node={node} edge={edge}/>
+            <ActionsBar />
             <Divider orientation="left" plain>Activities</Divider>
             <NodesTable/>
         </Drawer>
